requirements.txt
python-telegram-bot==20.3
SQLAlchemy==2.0.15
APScheduler==3.10.1
python-dotenv==1.0.0
psycopg2-binary==2.9.6
